[ calendar-UPDATE ] 이전/다음달 이동 버튼의 공통 부분을 custom-hooks으로 만들어라 - culturelist와 calendar컴포넌트에서 이전/다음달로 이동하는 함수를 사용하고 있었다. 이를 custom-hooks으로 생성하였다 - 오늘 날짜 데이터인 new Date를 커스텀훅인 useMonthNavigation의 인수로 보낸다. - 커스텀 훅 컴포넌트에서는 date를 받아와서 useState로 관리 및 이전/다음달의 데이터 정보를 업데이트하여 다시 return한다.
